[iOS] refactor: TableView Delegate 분리중
Issue #38
